requirements.txt
aiogram>=3.7.0
python-dotenv==1.0.0
aiosqlite==0.19.0
apscheduler==3.10.4
